Move fetch to newfetch.
It's not done yet and I want to give cmd/fetch its own package.
