+++ claat/newfetch/inprocessfetcher/inprocess.go
